[FIX] stock: avoid internal SM for inter-warehouses transfers
When using an internal transfer to move some products between two
warehouses, the forecasted inventory becomes incorrect

To reproduce the issue:
(Let WH01 be the existing warehouse)
1. In Settings, enable "Multi-Warehouses"
2. Create a second warehouse WH02
3. Create a storable product P
4. Update its quantity: 1 x P at WH01/Stock
5. Create a planned internal transfer (the warehouse does not matter):
    - Source: WH01/Stock
    - Destination: WH02/Stock
    - Operations:
        - 1 x P
6. Mark the transfer as Todo
7. Consult the Forecasted Inventory report:
    - Filters:
        - Forecasted Stock
        - Product: P
    - Group By:
        - Warehouse

Error: The line for WH02 is missing and the line for WH01 is incorrect
(qty is 1 for next days while it should be 0)

The report used (`report_stock_quantity`) does not handle this transfer.
Because both locations (source and destination) are defined, the SM is
not considered as an in-move or out-move:
https://github.com/odoo/odoo/blob/07f0ffad6ecc08f4165902db45ae96e4b41199e8/addons/stock/report/report_stock_quantity.py#L38-L43
Moreover, it would be hard to change the SQL view to generate two lines
(one 'in' and one 'out') from the same SM (the internal move).

The graph on the forecasted report of product P is also incorrect (it
uses `report_stock_quantity` to get the data, so if the user selects a
specific warehouse in the filters, the result will be incorrect)

To easy all computations, the user should rather use two pickings
(delivery+receipt) when moving some products between two warehouses.

OPW-2752017

closes odoo/odoo#86960

Signed-off-by: Steve Van Essche <svs@odoo.com>